03_sqlite_to_postgres/requirements.txt
flake8==4.0  # Разумеется, вам потребуется линтер :) 
psycopg2-binary==2.9
python-dotenv==0.20.0